adding the condition to block updates in launch_chffrplus.sh

+++ launch_chffrplus.sh
@@ -4,12 +4,17 @@ if [ -z "$PASSIVE" ]; then
   export PASSIVE="1"
 fi
 
+. /data/openpilot/selfdrive/car/tesla/readconfig.sh
+
+
 function launch {
   # apply update
-  if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
-     git reset --hard @{u} &&
-     git clean -xdf &&
-     exec "${BASH_SOURCE[0]}"
+  if [ $do_auto_update == "True" ]; then
+    if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
+      git reset --hard @{u} &&
+      git clean -xdf &&
+      exec "${BASH_SOURCE[0]}"
+    fi
   fi
   
   # no cpu rationing for now
